Use rules_license to provide licensing information for units
rules_license is a nascent ruleset that will become the de facto
standard for specifying license information in a bazel monorepo.

This commit pulls in rules_license and uses it to specify the
Apache-2.0 license that is used by au.

This required relocating the python targets out of the root build
file and into tools/docs. I needed to do this because defining the
license target in the root BUILD file would mean that consumers
of au would need to load the root build file (to get the license
target), which would mean that they'd need the `au_pip_deps` repo,
which is really only a development dependency.

Ideally, au's bazel releases (what is pulled in by other
bazel repos in their WORKSPACE file), would only package the things that
were needed and not any additional development files. Until then, clean
things up a bit so that repos that _do_ pull in `au` via a full release
tar.gz can get correct license information without needing to load au's
dev dependencies.

See:
- https://github.com/bazelbuild/bazel/discussions/22966